Moving a DTO class from Model to DTO dir to avoid autowiring error
