+++ app/bundles/MarketplaceBundle/Tests/Functional/Command/InstallCommandTest.php
@@ -9,21 +9,22 @@ use InvalidArgumentException;
 use Mautic\CoreBundle\Helper\ComposerHelper;
 use Mautic\CoreBundle\Test\AbstractMauticTestCase;
 use Mautic\MarketplaceBundle\Command\InstallCommand;
+use Mautic\MarketplaceBundle\DTO\ConsoleOutput;
 use Mautic\MarketplaceBundle\DTO\PackageDetail;
 use Mautic\MarketplaceBundle\Exception\ApiException;
-use Mautic\MarketplaceBundle\Model\ConsoleOutputModel;
 use Mautic\MarketplaceBundle\Model\PackageModel;
 use PHPUnit\Framework\Assert;
+use PHPUnit\Framework\MockObject\MockObject;
 
 final class InstallCommandTest extends AbstractMauticTestCase
 {
     /**
-     * @var \PHPUnit\Framework\MockObject\MockObject&ComposerHelper
+     * @var MockObject&ComposerHelper
      */
     private $composerHelper;
 
     /**
-     * @var \PHPUnit\Framework\MockObject\MockObject&PackageModel
+     * @var MockObject&PackageModel
      */
     private $packageModel;
 
@@ -45,7 +46,7 @@ final class InstallCommandTest extends AbstractMauticTestCase
 
         $this->composerHelper->method('install')
             ->with($this->packageName)
-            ->willReturn(new ConsoleOutputModel(0, 'OK'));
+            ->willReturn(new ConsoleOutput(0, 'OK'));
 
         $command = new InstallCommand($this->composerHelper, $this->packageModel);
 
@@ -66,7 +67,7 @@ final class InstallCommandTest extends AbstractMauticTestCase
 
         $this->composerHelper->method('install')
             ->with($this->packageName)
-            ->willReturn(new ConsoleOutputModel(0, 'OK'));
+            ->willReturn(new ConsoleOutput(0, 'OK'));
 
         $command = new InstallCommand($this->composerHelper, $this->packageModel);
 
@@ -145,7 +146,7 @@ final class InstallCommandTest extends AbstractMauticTestCase
 
         $this->composerHelper->method('install')
             ->with($packageName)
-            ->willReturn(new ConsoleOutputModel(1, 'Something went wrong during the installation'));
+            ->willReturn(new ConsoleOutput(1, 'Something went wrong during the installation'));
 
         $this->packageModel->method('getPackageDetail')
             ->with($packageName)
